MainMgmt_back/requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
celery==5.4.0
certifi==2024.7.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
Django==5.0.3
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==25.1
django-redis==5.4.0
django-shortuuidfield==0.1.3
django-timezone-field==7.1
djangorestframework==3.15.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
mysqlclient==2.2.7
numpy==2.3.4
packaging==25.0
pandas==2.2.2
pillow==10.4.0
prompt_toolkit==3.0.52
PyJWT==2.8.0
PyMySQL==1.1.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
rpds-py==0.27.1
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.14
